www/maxicourses_test/requirements.txt
playwright==1.46.0
playwright-stealth==1.0.6
python-dotenv==1.0.1
rich==13.7.1
flask==3.0.3
Pillow==10.4.0
